pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "pandas>=2.3.0",
    "pandas-stubs==2.2.3.250527",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.1",
    "uvicorn>=0.35.0",
]
